adjust hardware check
