+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "flux",
-  "version": "2.0.1",
+  "version": "2.0.2",
   "description": "Flux, Your Gateway to a Decentralized World",
   "repository": {
     "type": "git",
